--- mobile/app/src/main/java/br/edu/puc/superid/Models/Senha.kt
@@ -1,10 +1,0 @@
-package br.edu.puc.superid.Models
-
-data class Senha (
-    var guid: String,
-    var login: String,
-    var nomeCategoria: String,
-    var senha: String,
-    var accessToken: String,
-    var uidUsuario: String
-)
