+++ mobile/app/src/main/java/br/edu/puc/superid/database/FirestoreHandler.kt
@@ -4,7 +4,6 @@ import android.util.Log
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import android.util.Base64
-import br.edu.puc.superid.Models.Senha
 import br.edu.puc.superid.auth.AuthHandler
 import com.google.firebase.auth.FirebaseAuth
 import kotlinx.coroutines.tasks.await
@@ -13,6 +12,15 @@ import javax.crypto.spec.SecretKeySpec
 import java.security.SecureRandom
 import java.util.UUID
 
+data class Senha (
+    var nome: String,
+    var guid: String,
+    var login: String,
+    var nomeCategoria: String,
+    var senha: String,
+    var accessToken: String,
+    var uidUsuario: String
+)
 
 class FirestoreHandler {
     private val db = Firebase.firestore
@@ -70,7 +78,7 @@ class FirestoreHandler {
         return UUID.randomUUID().toString()
     }
 
-    fun cadastrarSenha(login: String?, categoria: String, senha: String) {
+    fun cadastrarSenha(nome: String, login: String?, categoria: String, senha: String) {
         val guid = gerarGuid()
 
         val user = FirebaseAuth.getInstance().currentUser
@@ -80,6 +88,7 @@ class FirestoreHandler {
         val senhaCriptografada = criptografarSenha(senha)
 
         val senhaData = hashMapOf(
+            "nome" to nome,
             "guid" to guid,
             "login" to login,
             "nomeCategoria" to categoria,
@@ -134,6 +143,7 @@ class FirestoreHandler {
             .addOnCompleteListener { task ->
                 task.addOnSuccessListener { documents ->
                     for (document in documents) {
+                        val nome = document.getString("nome").toString()
                         val guid = document.getString("guid").toString()
                         val login = document.getString("login").toString()
                         val nomeCategoria = document.getString("nomeCategoria").toString()
@@ -141,7 +151,7 @@ class FirestoreHandler {
                         val accessToken = document.getString("accessToken").toString()
                         val uidUsuario = document.getString("uidUsuario").toString()
 
-                        var senhaData = Senha(guid, login, nomeCategoria, senha, accessToken, uidUsuario)
+                        var senhaData = Senha(nome, guid, login, nomeCategoria, senha, accessToken, uidUsuario)
 
                         listaDeSenhas.add(senhaData)
                     }
@@ -163,6 +173,7 @@ class FirestoreHandler {
             .addOnCompleteListener { task ->
                 task.addOnSuccessListener { documents ->
                     for (document in documents) {
+                        val nome = document.getString("nome").toString()
                         val guid = document.getString("guid").toString()
                         val login = document.getString("login").toString()
                         val nomeCategoria = document.getString("nomeCategoria").toString()
@@ -170,7 +181,7 @@ class FirestoreHandler {
                         val accessToken = document.getString("accessToken").toString()
                         val uidUsuario = document.getString("uidUsuario").toString()
 
-                        var senhaData = Senha(guid, login, nomeCategoria, senha, accessToken, uidUsuario)
+                        var senhaData = Senha(nome, guid, login, nomeCategoria, senha, accessToken, uidUsuario)
 
                         listaDeSenhas.add(senhaData)
                     }
